# common/api/content/v1/comment_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v6.33.1
// source: content/v1/comment.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationContentCommentServiceAdd = "/common.api.content.v1.ContentCommentService/Add"
const OperationContentCommentServiceLike = "/common.api.content.v1.ContentCommentService/Like"
const OperationContentCommentServicePage = "/common.api.content.v1.ContentCommentService/Page"
const OperationContentCommentServiceThank = "/common.api.content.v1.ContentCommentService/Thank"
const OperationContentCommentServiceUpdateStatus = "/common.api.content.v1.ContentCommentService/UpdateStatus"

type ContentCommentServiceHTTPServer interface {
	// Add 新增评论
	Add(context.Context, *AddCommentRequest) (*AddCommentReply, error)
	// Like 点赞评论
	Like(context.Context, *LikeCommentRequest) (*LikeCommentReply, error)
	// Page 查询评论列表
	Page(context.Context, *PageCommentRequest) (*PageCommentReply, error)
	// Thank 感谢评论
	Thank(context.Context, *ThankCommentRequest) (*ThankCommentReply, error)
	// UpdateStatus 修改评论状态
	UpdateStatus(context.Context, *UpdateStatusCommentRequest) (*UpdateStatusCommentReply, error)
}

func RegisterContentCommentServiceHTTPServer(s *http.Server, srv ContentCommentServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/comment/add", _ContentCommentService_Add1_HTTP_Handler(srv))
	r.POST("/v1/comment/page", _ContentCommentService_Page3_HTTP_Handler(srv))
	r.POST("/v1/comment/like", _ContentCommentService_Like1_HTTP_Handler(srv))
	r.POST("/v1/comment/thank", _ContentCommentService_Thank1_HTTP_Handler(srv))
	r.POST("/v1/comment/updateStatus", _ContentCommentService_UpdateStatus0_HTTP_Handler(srv))
}

func _ContentCommentService_Add1_HTTP_Handler(srv ContentCommentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AddCommentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentCommentServiceAdd)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Add(ctx, req.(*AddCommentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AddCommentReply)
		return ctx.Result(200, reply)
	}
}

func _ContentCommentService_Page3_HTTP_Handler(srv ContentCommentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PageCommentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentCommentServicePage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Page(ctx, req.(*PageCommentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PageCommentReply)
		return ctx.Result(200, reply)
	}
}

func _ContentCommentService_Like1_HTTP_Handler(srv ContentCommentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LikeCommentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentCommentServiceLike)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Like(ctx, req.(*LikeCommentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LikeCommentReply)
		return ctx.Result(200, reply)
	}
}

func _ContentCommentService_Thank1_HTTP_Handler(srv ContentCommentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ThankCommentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentCommentServiceThank)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Thank(ctx, req.(*ThankCommentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ThankCommentReply)
		return ctx.Result(200, reply)
	}
}

func _ContentCommentService_UpdateStatus0_HTTP_Handler(srv ContentCommentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateStatusCommentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentCommentServiceUpdateStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateStatus(ctx, req.(*UpdateStatusCommentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateStatusCommentReply)
		return ctx.Result(200, reply)
	}
}

type ContentCommentServiceHTTPClient interface {
	// Add 新增评论
	Add(ctx context.Context, req *AddCommentRequest, opts ...http.CallOption) (rsp *AddCommentReply, err error)
	// Like 点赞评论
	Like(ctx context.Context, req *LikeCommentRequest, opts ...http.CallOption) (rsp *LikeCommentReply, err error)
	// Page 查询评论列表
	Page(ctx context.Context, req *PageCommentRequest, opts ...http.CallOption) (rsp *PageCommentReply, err error)
	// Thank 感谢评论
	Thank(ctx context.Context, req *ThankCommentRequest, opts ...http.CallOption) (rsp *ThankCommentReply, err error)
	// UpdateStatus 修改评论状态
	UpdateStatus(ctx context.Context, req *UpdateStatusCommentRequest, opts ...http.CallOption) (rsp *UpdateStatusCommentReply, err error)
}

type ContentCommentServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewContentCommentServiceHTTPClient(client *http.Client) ContentCommentServiceHTTPClient {
	return &ContentCommentServiceHTTPClientImpl{client}
}

// Add 新增评论
func (c *ContentCommentServiceHTTPClientImpl) Add(ctx context.Context, in *AddCommentRequest, opts ...http.CallOption) (*AddCommentReply, error) {
	var out AddCommentReply
	pattern := "/v1/comment/add"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentCommentServiceAdd))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Like 点赞评论
func (c *ContentCommentServiceHTTPClientImpl) Like(ctx context.Context, in *LikeCommentRequest, opts ...http.CallOption) (*LikeCommentReply, error) {
	var out LikeCommentReply
	pattern := "/v1/comment/like"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentCommentServiceLike))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Page 查询评论列表
func (c *ContentCommentServiceHTTPClientImpl) Page(ctx context.Context, in *PageCommentRequest, opts ...http.CallOption) (*PageCommentReply, error) {
	var out PageCommentReply
	pattern := "/v1/comment/page"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentCommentServicePage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Thank 感谢评论
func (c *ContentCommentServiceHTTPClientImpl) Thank(ctx context.Context, in *ThankCommentRequest, opts ...http.CallOption) (*ThankCommentReply, error) {
	var out ThankCommentReply
	pattern := "/v1/comment/thank"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentCommentServiceThank))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateStatus 修改评论状态
func (c *ContentCommentServiceHTTPClientImpl) UpdateStatus(ctx context.Context, in *UpdateStatusCommentRequest, opts ...http.CallOption) (*UpdateStatusCommentReply, error) {
	var out UpdateStatusCommentReply
	pattern := "/v1/comment/updateStatus"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentCommentServiceUpdateStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}