# common/api/content/v1/domain_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v6.33.1
// source: content/v1/domain.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationContentDomainServiceAdds = "/common.api.content.v1.ContentDomainService/Adds"
const OperationContentDomainServicePage = "/common.api.content.v1.ContentDomainService/Page"
const OperationContentDomainServiceUpdate = "/common.api.content.v1.ContentDomainService/Update"

type ContentDomainServiceHTTPServer interface {
	// Adds 新增领域
	Adds(context.Context, *AddDomainsRequest) (*AddDomainsReply, error)
	// Page 查询领域
	Page(context.Context, *PageDomainRequest) (*PageDomainReply, error)
	// Update 修改领域信息
	Update(context.Context, *UpdateDomainRequest) (*UpdateDomainReply, error)
}

func RegisterContentDomainServiceHTTPServer(s *http.Server, srv ContentDomainServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/domain/adds", _ContentDomainService_Adds1_HTTP_Handler(srv))
	r.POST("/v1/domain/update", _ContentDomainService_Update2_HTTP_Handler(srv))
	r.POST("/v1/domain/page", _ContentDomainService_Page4_HTTP_Handler(srv))
}

func _ContentDomainService_Adds1_HTTP_Handler(srv ContentDomainServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AddDomainsRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentDomainServiceAdds)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Adds(ctx, req.(*AddDomainsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AddDomainsReply)
		return ctx.Result(200, reply)
	}
}

func _ContentDomainService_Update2_HTTP_Handler(srv ContentDomainServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateDomainRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentDomainServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*UpdateDomainRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateDomainReply)
		return ctx.Result(200, reply)
	}
}

func _ContentDomainService_Page4_HTTP_Handler(srv ContentDomainServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PageDomainRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationContentDomainServicePage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Page(ctx, req.(*PageDomainRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PageDomainReply)
		return ctx.Result(200, reply)
	}
}

type ContentDomainServiceHTTPClient interface {
	// Adds 新增领域
	Adds(ctx context.Context, req *AddDomainsRequest, opts ...http.CallOption) (rsp *AddDomainsReply, err error)
	// Page 查询领域
	Page(ctx context.Context, req *PageDomainRequest, opts ...http.CallOption) (rsp *PageDomainReply, err error)
	// Update 修改领域信息
	Update(ctx context.Context, req *UpdateDomainRequest, opts ...http.CallOption) (rsp *UpdateDomainReply, err error)
}

type ContentDomainServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewContentDomainServiceHTTPClient(client *http.Client) ContentDomainServiceHTTPClient {
	return &ContentDomainServiceHTTPClientImpl{client}
}

// Adds 新增领域
func (c *ContentDomainServiceHTTPClientImpl) Adds(ctx context.Context, in *AddDomainsRequest, opts ...http.CallOption) (*AddDomainsReply, error) {
	var out AddDomainsReply
	pattern := "/v1/domain/adds"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentDomainServiceAdds))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Page 查询领域
func (c *ContentDomainServiceHTTPClientImpl) Page(ctx context.Context, in *PageDomainRequest, opts ...http.CallOption) (*PageDomainReply, error) {
	var out PageDomainReply
	pattern := "/v1/domain/page"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentDomainServicePage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Update 修改领域信息
func (c *ContentDomainServiceHTTPClientImpl) Update(ctx context.Context, in *UpdateDomainRequest, opts ...http.CallOption) (*UpdateDomainReply, error) {
	var out UpdateDomainReply
	pattern := "/v1/domain/update"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationContentDomainServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}