# common/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/comment/add:
        post:
            tags:
                - ContentCommentService
            description: 新增评论
            operationId: ContentCommentService_Add
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.AddCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.AddCommentReply'
    /v1/comment/like:
        post:
            tags:
                - ContentCommentService
            description: 点赞评论
            operationId: ContentCommentService_Like
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.LikeCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.LikeCommentReply'
    /v1/comment/page:
        post:
            tags:
                - ContentCommentService
            description: 查询评论列表
            operationId: ContentCommentService_Page
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.PageCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.PageCommentReply'
    /v1/comment/thank:
        post:
            tags:
                - ContentCommentService
            description: 感谢评论
            operationId: ContentCommentService_Thank
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.ThankCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.ThankCommentReply'
    /v1/comment/updateStatus:
        post:
            tags:
                - ContentCommentService
            description: 修改评论状态
            operationId: ContentCommentService_UpdateStatus
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.UpdateStatusCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.UpdateStatusCommentReply'
    /v1/domain/adds:
        post:
            tags:
                - ContentDomainService
            description: 新增领域
            operationId: ContentDomainService_Adds
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.AddDomainsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.AddDomainsReply'
    /v1/domain/page:
        post:
            tags:
                - ContentDomainService
            description: 查询领域
            operationId: ContentDomainService_Page
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.PageDomainRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.PageDomainReply'
    /v1/domain/update:
        post:
            tags:
                - ContentDomainService
            description: 修改领域信息
            operationId: ContentDomainService_Update
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.UpdateDomainRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.UpdateDomainReply'
...
components:
    schemas:
...
        common.api.common.v1.Int32Range:
            type: object
            properties:
                start:
                    type: integer
                    description: 开始
                    format: int32
                end:
                    type: integer
                    description: 结束
                    format: int32
        common.api.common.v1.PageReply:
            type: object
            properties:
                total:
                    type: integer
                    description: 总数
                    format: uint32
                page:
                    type: integer
                    description: 页码
                    format: uint32
                size:
                    type: integer
                    description: 页大小
                    format: uint32
        common.api.common.v1.PageRequest:
            type: object
            properties:
                page:
                    type: integer
                    description: 页码
                    format: uint32
                size:
                    type: integer
                    description: 页大小
                    format: uint32
...
        common.api.content.v1.AddCommentReply:
            type: object
            properties:
                comment:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.Comment'
                    description: 评论
        common.api.content.v1.AddCommentRequest:
            type: object
            properties:
                article_id:
                    type: string
                    description: 文章Id
                content:
                    type: string
                    description: 评论内容
                reply_id:
                    type: string
                    description: 回复评论Id
        common.api.content.v1.AddDomainsReply:
            type: object
            properties:
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/common.api.content.v1.Domain'
                    description: 添加的领域信息
        common.api.content.v1.AddDomainsRequest:
            type: object
            properties:
                domains:
                    type: array
                    items:
                        $ref: '#/components/schemas/common.api.content.v1.DomainSave'
                    description: 领域信息
...
        common.api.content.v1.Comment:
            type: object
            properties:
                created_at:
                    type: string
                    description: 创建时间
                    format: date-time
                updated_at:
                    type: string
                    description: 更新时间
                    format: date-time
                created_by:
                    type: string
                    description: 创建人
                updated_by:
                    type: string
                    description: 更新人
                id:
                    type: string
                    description: 评论Id
                article_id:
                    type: string
                    description: 文章Id
                content:
                    type: string
                    description: 评论内容
                content_render:
                    type: string
                    description: 评论内容渲染
                level:
                    type: integer
                    description: 评论层级
                    format: int32
                parent_id:
                    type: string
                    description: 父级评论Id
                reply_id:
                    type: string
                    description: 回复评论Id
                status:
                    type: integer
                    description: 状态
                    format: int32
                thank_count:
                    type: integer
                    description: 感谢数量
                    format: int32
                like_count:
                    type: integer
                    description: 点赞数量
                    format: int32
                collect_count:
                    type: integer
                    description: 收藏数量
                    format: int32
                reply_count:
                    type: integer
                    description: 子评论数量
                    format: int32
                user:
                    allOf:
                        - $ref: '#/components/schemas/common.api.user.v1.User'
                    description: 用户信息
                reply_user:
                    allOf:
                        - $ref: '#/components/schemas/common.api.user.v1.User'
                    description: 回复评论的用户信息
        common.api.content.v1.CommentQueryParams:
            type: object
            properties:
                comment_id:
                    type: string
                    description: 评论Id
                article_id:
                    type: string
                    description: 文章Id
                parent_id:
                    type: string
                    description: 父级评论Id
                reply_id:
                    type: string
                    description: 回复评论Id
                order:
                    type: integer
                    description: 排序 0-最新 1-最热
                    format: int32
                user_id:
                    type: string
                    description: 用户Id
                level:
                    type: integer
                    description: 层级
                    format: int32
...
        common.api.content.v1.Domain:
            type: object
            properties:
                created_at:
                    type: string
                    description: 创建时间
                    format: date-time
                updated_at:
                    type: string
                    description: 更新时间
                    format: date-time
                created_by:
                    type: string
                    description: 创建人
                updated_by:
                    type: string
                    description: 更新人
                id:
                    type: string
                    description: 主键
                name:
                    type: string
                    description: 领域名称
                description:
                    type: string
                    description: 描述
                status:
                    type: integer
                    description: 状态 0-正常，1-禁用
                    format: int32
                url:
                    type: string
                    description: 链接
                icon:
                    type: string
                    description: 图标
                tagCount:
                    type: integer
                    description: 标签数量
                    format: int32
                isNav:
                    type: boolean
                    description: 是否导航
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/common.api.content.v1.Tag'
                    description: tags
        common.api.content.v1.DomainQueryParams:
            type: object
            properties:
                ids:
                    type: array
                    items:
                        type: string
                    description: 主键
                name:
                    type: string
                    description: 领域名称
                description:
                    type: string
                    description: 描述
                status:
                    type: string
                    description: 状态 0-正常，1-禁用
                url:
                    type: string
                    description: 链接
                icon:
                    type: string
                    description: 图标
                isNav:
                    type: boolean
                    description: 是否导航
                tagCount:
                    allOf:
                        - $ref: '#/components/schemas/common.api.common.v1.Int32Range'
                    description: 标签数量范围
        common.api.content.v1.DomainSave:
            type: object
            properties:
                id:
                    type: string
                    description: 主键
                name:
                    type: string
                    description: 领域名称
                description:
                    type: string
                    description: 描述
                status:
                    type: integer
                    description: 状态 0-正常，1-禁用
                    format: int32
                url:
                    type: string
                    description: 链接
                icon:
                    type: string
                    description: 图标
                isNav:
                    type: boolean
                    description: 是否导航
...
        common.api.content.v1.LikeCommentReply:
            type: object
            properties: {}
        common.api.content.v1.LikeCommentRequest:
            type: object
            properties:
                id:
                    type: string
                    description: 评论Id
                active:
                    type: boolean
                    description: 是否点赞
...
        common.api.content.v1.PageCommentReply:
            type: object
            properties:
                page:
                    allOf:
                        - $ref: '#/components/schemas/common.api.common.v1.PageReply'
                    description: 分页
                rows:
                    type: array
                    items:
                        $ref: '#/components/schemas/common.api.content.v1.Comment'
                    description: 评论列表
        common.api.content.v1.PageCommentRequest:
            type: object
            properties:
                page:
                    allOf:
                        - $ref: '#/components/schemas/common.api.common.v1.PageRequest'
                    description: 分页
                query:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.CommentQueryParams'
                    description: 查询参数
        common.api.content.v1.PageDomainReply:
            type: object
            properties:
                page:
                    allOf:
                        - $ref: '#/components/schemas/common.api.common.v1.PageReply'
                    description: 分页信息
                rows:
                    type: array
                    items:
                        $ref: '#/components/schemas/common.api.content.v1.Domain'
                    description: 领域信息
        common.api.content.v1.PageDomainRequest:
            type: object
            properties:
                page:
                    allOf:
                        - $ref: '#/components/schemas/common.api.common.v1.PageRequest'
                    description: 分页参数
                query:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.DomainQueryParams'
                    description: 查询参数
...
        common.api.content.v1.Tag:
            type: object
            properties:
                created_at:
                    type: string
                    description: 创建时间
                    format: date-time
                updated_at:
                    type: string
                    description: 更新时间
                    format: date-time
                created_by:
                    type: string
                    description: 创建人
                updated_by:
                    type: string
                    description: 更新人
                id:
                    type: string
                    description: 主键
                name:
                    type: string
                    description: 标签名称
                description:
                    type: string
                    description: 描述
                domain_id:
                    type: string
                    description: 所属领域Id
                status:
                    type: integer
                    description: 标签状态：0-正常，1-禁用
                    format: int32
                article_count:
                    type: integer
                    description: 文章数
                    format: int32
...
        common.api.content.v1.ThankCommentReply:
            type: object
            properties: {}
        common.api.content.v1.ThankCommentRequest:
            type: object
            properties:
                id:
                    type: string
                    description: 评论Id
                active:
                    type: boolean
                    description: 是否感谢
...
        common.api.content.v1.UpdateDomainReply:
            type: object
            properties:
                data:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.Domain'
                    description: 修改的领域信息
        common.api.content.v1.UpdateDomainRequest:
            type: object
            properties:
                domain:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.DomainSave'
                    description: 领域信息
        common.api.content.v1.UpdateStatusCommentReply:
            type: object
            properties: {}
        common.api.content.v1.UpdateStatusCommentRequest:
            type: object
            properties:
                id:
                    type: string
                    description: 评论Id
                status:
                    type: string
                    description: 状态
...